rootext/calibration: Unify into header only library
Convert the Calibration.{hh,cc} rootext library into a header only library.
This is done to simplify the loading of the Calibration class via the
ROOT CLING Interpreter later on.

This patch also reworks some of the functions, but does not change any
behavior.

Updates the license header information to a SPDX identifier format.
